HOSDict: move to dict folder
